Add support for hardware interrupts
Added support for hardware interrupts using a driver for two chained 8259 PICs.
Added two new entries in the IDT - a handler for the Keyboard and Timer
interrupt.
Inserted an interrupt-free context in both _print functions in order
to prevent deadlocks.
Added a HLT loop in several places to reduce the Kernel's CPU usage when
there is no work to do.