Narrow an unsafe block
Made the init_pics function safe again by moving its body
in an unsafe block. This is OK since the unsafeness lies in
the offset values passed to the ChainedPics constructor.
The init_pics functions is supposed to be safe, provided that
the correct values are passed to the constructor.